Move the endpointslices structure to a different package
This is required to avoid circular dependencies between the k8s package
and the new added controllers package.
